Relocate WorkerRole/OcrBackend to a neutral providers.roles module
These two identifiers outlive the in-process worker IPC layer: the fleet
and every surface address roles by them, and OcrBackend names the OCR
path. Move them out of providers.worker.transport ahead of deleting that
package, and repoint the eleven external import sites.
